TelegramBotWithllama3.2/requirements.txt
aiogram
openai==0.28
python-dotenv
ollama